+++ source/components/PasswordInput.js
@@ -1,98 +1,151 @@
 // @flow
-import React, { Component } from 'react';
-// $FlowFixMe
-import type { ComponentType, Element } from 'react';
+import stringEntropy from 'fast-password-entropy';
+import React, { StatelessFunctionalComponent, useContext } from 'react';
+import { ThemeContext } from './HOC/ThemeContext';
 
 // internal utility functions
-import { createEmptyContext, withTheme } from './HOC/withTheme';
 import { composeTheme, addThemeId } from '../utils/themes';
 
 // import constants
 import { IDENTIFIERS } from '.';
-import type { ThemeContextProp } from './HOC/withTheme';
+import type { InputProps } from './Input';
 
-export type PasswordInputProps = {
-  autoFocus?: boolean,
-  className?: string,
-  context: ThemeContextProp,
-  disabled?: boolean,
-  error?: string,
-  label?: string | Element<any>,
+export const calculatePasswordScore = (
+  password: string,
+  entropyFactor: number = 0.01
+): number => Math.min((stringEntropy(password) * entropyFactor).toFixed(2), 1);
+
+const STATE = {
+  DEFAULT: 'default',
+  ERROR: 'error',
+  INSECURE: 'insecure',
+  WEAK: 'weak',
+  STRONG: 'strong',
+};
+
+export type PasswordInputProps = InputProps & {
+  debounceDelay?: number,
+  entropyFactor?: number,
+  isPasswordRepeat?: boolean,
+  isShowingTooltipOnFocus: boolean,
+  isShowingTooltipOnHover: boolean,
   isTooltipOpen: boolean,
-  onBlur?: Function,
-  onChange?: Function,
-  onFocus?: Function,
-  placeholder?: string,
-  readOnly?: boolean,
-  skin?: ComponentType<any>,
-  theme: ?Object,
-  themeId: string,
-  themeOverrides: Object,
-  value: ?string,
-  tooltip?: string,
-  score?: number,
-  state?: $Values<typeof PasswordInputBase.STATE>,
+  minLength?: number,
+  minStrongScore?: number,
+  repeatPassword?: string,
+  state?: $Values<typeof STATE>,
+  passwordFeedbacks?: {
+    insecure: string,
+    weak: string,
+    strong: string,
+    noMatch: string,
+  },
+  tooltip?: string | boolean,
+  useDebounce?: boolean,
 };
 
-type State = {
-  composedTheme: Object,
+export const PasswordInput: StatelessFunctionalComponent<PasswordInputProps> = (
+  props
+) => {
+  const {
+    context,
+    passwordFeedbacks,
+    entropyFactor,
+    error,
+    minLength,
+    minStrongScore,
+    isPasswordRepeat,
+    repeatPassword,
+    skin,
+    state,
+    theme,
+    themeOverrides,
+    tooltip,
+    ...rest
+  } = props;
+
+  // Theme
+  const themeContext = context || useContext(ThemeContext);
+  const composedTheme = composeTheme(
+    addThemeId(theme || themeContext.theme, props.themeId),
+    addThemeId(props.themeOverrides, props.themeId),
+    themeContext.ROOT_THEME_API
+  );
+  // Skin
+  const PasswordInputSkin =
+    skin || themeContext.skins[IDENTIFIERS.PASSWORD_INPUT];
+
+  // Logic
+  let dynamicState = PasswordInput.STATE.DEFAULT;
+  let passwordFeedback = null;
+  const password = props.value;
+  const score = calculatePasswordScore(password, entropyFactor);
+  const isValidPassword = password.length >= minLength;
+  const isNotEmpty = password.length > 0;
+  const isRepeat = !!repeatPassword;
+
+  if (error) {
+    dynamicState = PasswordInput.STATE.ERROR;
+    passwordFeedback = error;
+  } else if (tooltip) {
+    passwordFeedback = tooltip;
+  } else if (isPasswordRepeat) {
+    if (repeatPassword === props.value) {
+      dynamicState = PasswordInput.STATE.DEFAULT;
+      passwordFeedback = null;
+    } else {
+      dynamicState = PasswordInput.STATE.ERROR;
+      passwordFeedback = passwordFeedbacks.noMatch;
+    }
+  } else if (isValidPassword) {
+    if (score < minStrongScore) {
+      dynamicState = PasswordInput.STATE.WEAK;
+      passwordFeedback = passwordFeedbacks[PasswordInput.STATE.WEAK];
+    } else {
+      dynamicState = PasswordInput.STATE.STRONG;
+      passwordFeedback = passwordFeedbacks[PasswordInput.STATE.STRONG];
+    }
+  } else if (isNotEmpty) {
+    dynamicState = PasswordInput.STATE.INSECURE;
+    passwordFeedback = passwordFeedbacks[PasswordInput.STATE.INSECURE];
+  }
+  return (
+    <PasswordInputSkin
+      error={error}
+      theme={composedTheme}
+      score={score}
+      state={state || dynamicState}
+      tooltip={passwordFeedback}
+      {...rest}
+    />
+  );
 };
 
-class PasswordInputBase extends Component<PasswordInputProps, State> {
+// Static Properties
 
-  static displayName = 'PasswordInput';
-  static STATE = {
-    DEFAULT: 'default',
-    ERROR: 'error',
-    WARNING: 'warning',
-    SUCCESS: 'success',
-  };
+PasswordInput.STATE = STATE;
 
-  static defaultProps = {
-    context: createEmptyContext(),
-    isTooltipOpen: false,
-    readOnly: false,
-    theme: null,
-    themeId: IDENTIFIERS.PASSWORD_INPUT,
-    themeOverrides: {},
-    value: null,
-    score: 0,
-    state: PasswordInputBase.STATE.DEFAULT,
-  };
+PasswordInput.displayName = 'PasswordInput';
 
-  constructor(props: PasswordInputProps) {
-    super(props);
-    const { context, themeId, theme, themeOverrides } = props;
-    this.state = {
-      composedTheme: composeTheme(
-        addThemeId(theme || context.theme, themeId),
-        addThemeId(themeOverrides, themeId),
-        context.ROOT_THEME_API
-      ),
-    };
-  }
-
-  render() {
-    // destructuring props ensures only the "...rest" get passed down
-    const {
-      context,
-      skin,
-      theme,
-      themeOverrides,
-      ...rest
-    } = this.props;
-
-    const PasswordInputSkin = skin || context.skins[IDENTIFIERS.PASSWORD_INPUT];
-
-    return (
-      <PasswordInputSkin
-        theme={this.state.composedTheme}
-        {...rest}
-      />
-    );
-  }
-}
-
-export const PasswordInput = withTheme(PasswordInputBase);
-
-PasswordInput.STATE = PasswordInputBase.STATE;
+PasswordInput.defaultProps = {
+  debounceDelay: 1000,
+  entropyFactor: 0.01,
+  passwordFeedbacks: {
+    insecure: 'insecure',
+    weak: 'weak',
+    strong: 'strong',
+    noMatch: "doesn't match",
+  },
+  isPasswordRepeat: false,
+  isTooltipOpen: false,
+  isShowingTooltipOnFocus: true,
+  isShowingTooltipOnHover: true,
+  minLength: 10,
+  minStrongScore: 0.75,
+  readOnly: false,
+  theme: null,
+  themeId: IDENTIFIERS.PASSWORD_INPUT,
+  themeOverrides: {},
+  useDebounce: true,
+  value: '',
+};