+++ source/skins/simple/InputSkin.js
@@ -8,6 +8,7 @@ import { isFunction } from 'lodash';
 
 // components
 import { FormField } from '../../components/FormField';
+import type { InputProps } from '../../components/Input';
 
 // skins
 import { FormFieldSkin } from './FormFieldSkin';
@@ -15,23 +16,10 @@ import { FormFieldSkin } from './FormFieldSkin';
 // internal utility functions
 import { pickDOMProps } from '../../utils/props';
 
-type Props = {
-  className?: ?string,
-  disabled?: boolean,
-  error?: string,
-  label?: string | Element<any>,
+type Props = InputProps & {
   inputRef: ElementRef<'input'>,
-  onBlur?: Function,
-  onChange?: Function,
-  onFocus?: Function,
-  onKeyPress?: Function,
-  placeholder?: string,
-  readOnly?: boolean,
-  selectedOption?: any,
   theme: Object,
   themeId: string,
-  value: string,
-  selectionRenderer?: Function,
 };
 
 export const InputSkin = (props: Props) => {
@@ -42,13 +30,15 @@ export const InputSkin = (props: Props) => {
       className={classnames([
         props.theme[props.themeId].input,
         props.disabled ? props.theme[props.themeId].disabled : null,
-        props.error ? props.theme[props.themeId].errored : null,
+        props.error || props.showErrorState
+          ? props.theme[props.themeId].errored
+          : null,
       ])}
       readOnly={props.readOnly}
     />
   );
 
-  const useSelectionRenderer = option => (
+  const useSelectionRenderer = (option) => (
     <div className={props.theme[props.themeId].customValueWrapper}>
       {renderInput()}
       <div className={props.theme[props.themeId].customValueBlock}>
@@ -59,7 +49,8 @@ export const InputSkin = (props: Props) => {
 
   const render = () => {
     // check if user has passed render prop "selectionRenderer"
-    const hasSelectionRenderer = props.selectionRenderer && isFunction(props.selectionRenderer);
+    const hasSelectionRenderer =
+      props.selectionRenderer && isFunction(props.selectionRenderer);
     if (hasSelectionRenderer) {
       return useSelectionRenderer(props.selectedOption);
     }